test: split eval corpus suites across files for CI shard balance
corpus-v0.3.test.ts (196s, 13 tasks) and corpus-v0.2.test.ts (65s,
5 tasks) each ran as one file, capping per-file CI sharding. Partition
the task lists into files of roughly equal measured cost; the per-task
runner/scoring scaffolding moves unchanged into the shared helper
eval/corpusExpertSuite.ts (same beforeAll CI guard, same beforeEach
tmpdir, same runTask inputs, same assertions).

- corpus-v0.3.test.ts: tasks 1-4 (29.5s)
- corpus-v0.3b.test.ts: tasks 5-8 (27.6s)
- corpus-v0.3c.test.ts: tasks 9-11 (23.0s)
- corpus-v0.3d.test.ts: tasks 12-13 (21.1s)
- corpus-v0.2.test.ts: tasks 1-3 (19.6s)
- corpus-v0.2.labeled.test.ts: labeled-* tasks 4-5 (12.8s)

No dedup performed (each task is an independent corpus entry). Test
counts unchanged: 13 -> 4+4+3+2, 5 -> 3+2.
